% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{datestamp}
\alias{datestamp}
\title{Datestamp character string}
\usage{
datestamp(filenameOrSomething, location = c("beforeFileExt", "start",
  "end"), sep = ".", dateFormat = "\%Y\%m\%d_\%H\%M")
}
\arguments{
\item{filenameOrSomething}{character string to be datestamped}

\item{location}{c("beforeFileExt", "start", "end"). where the datastamp should be inserted. 
default is before the file extension ("beforeFileExt")}

\item{sep}{character that separates the datastamp and the original characters. 
default = "."}

\item{dateFormat}{character formatting string. see \code{\link{format.POSIXct}} for more info.}
}
\value{
datastamped character string
}
\description{
Add a datestamp to a character string, with considerations for filenames
}
\examples{
datestamp("salad.pdf")    # "salad.20181211_1005.pdf"
datestamp("sample", location="end", sep=" ", dateFormat = "at \%H\%M")    #"sample at 1005"

}
