% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinformatic.R
\name{ensembl.to.hgnc}
\alias{ensembl.to.hgnc}
\title{Returns a dataframe of matched HGNC IDs for the inputted Ensembl IDs.}
\usage{
ensembl.to.hgnc(ensemblIDs, version = NULL, additional = c())
}
\arguments{
\item{ensemblIDs}{vector of character strings representing Ensembl IDs}

\item{version}{NULL for the newest version, or numerical Ensembl version}

\item{additional}{vector of additional biomaRt attributes (eg: "transcript_length"). See biomaRt::getBM().}
}
\value{
dataframe of "ensembl_gene_id", "hgnc_symbol" and any additional attributes
}
\description{
Returns a dataframe of matched HGNC IDs for the inputted Ensembl IDs.
}
