% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{png.date}
\alias{png.date}
\title{\code{\link{png}} wrapper with datestamp}
\usage{
png.date(filename = "Rplot\%03d.png", dateFormat = "\%Y\%m\%d_\%H\%M",
  width = 10, height = 10, units = "in", pointsize = 12,
  bg = "white", res = 300, type = c("cairo", "cairo-png", "Xlib",
  "quartz"), antialias = "none")
}
\description{
A wrapper of the \code{\link{png}} function that automatically datestamps the filename,
and uses a default of 10 inches by 10 inches at 300 dpi for image size
}
