% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{setwd.multi}
\alias{setwd.multi}
\title{Try \code{\link{setwd}} with multiple files}
\usage{
setwd.multi(multiPath)
}
\arguments{
\item{multiPath}{vector of characters representing inputs to \code{\link{setwd}}}
}
\value{
None
}
\description{
\code{\link{setwd}}s the first dir that exists in a vector of dir
}
\examples{
setwd.multi("path/on/hpc", "path/on/local")

}
