% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{source.multi}
\alias{source.multi}
\title{Try \code{\link{source}} with multiple files}
\usage{
source.multi(multiPath)
}
\arguments{
\item{multiPath}{vector of character strings representing inputs to \code{\link{source}}}
}
\value{
None
}
\description{
\code{\link{source}}s the first file that exists in a vector of files.
}
\examples{
source.multi(c("path/on/hpc", "path/on/local"))

}
